Refactored Code!
